Moves crucial Storage parts into seperated Files
new subfolder for engines and moved flatfile to location
table, column, database and tablemetadata are now located in meta.rs
some work in query to accept the new changes to column
same for table_test
new file types for future use of SqlTypes and more
