+++ server/src/storage/engine/mod.rs
@@ -1,0 +1,3 @@
+mod flatfile;
+
+pub use self::flatfile::FlatFile;
